% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_log_2.R
\name{get_log_file}
\alias{get_log_file}
\title{write_log_2
Generate a timestamped log file path inside the provided configuration folder.}
\usage{
get_log_file(config_folder)
}
\arguments{
\item{config_folder}{str. Path to the folder containing the analysis configuration.}
}
\value{
str Full file path to a new log file, named as analysis_<timestamp>.log.
}
\description{
write_log_2
Generate a timestamped log file path inside the provided configuration folder.
}
