% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_log_message_2.R
\name{write_log_2}
\alias{write_log_2}
\title{add_log_message_2}
\usage{
write_log_2(message, log_file_2)
}
\arguments{
\item{message}{The actual commit message before push}

\item{log_file_2}{The path of the downstream analysis log file}
}
\description{
Log a message with a time stamp.
}
\keyword{internal}
